fixed scripts build, clean and added script for rebuild

+++ project/build_coex.sh
@@ -1,7 +1,9 @@
 cd computer_examination
 QMAKE=""
 
-rm ../coex
+if [ -f "../coex" ]; then
+  rm ../coex
+fi
 
 # usr/bin/qmake-qt4
 if [ -f "/usr/bin/qmake-qt4" ]; then
@@ -13,7 +15,7 @@ if [ -f "/bin/qmake-qt4" ]; then
   QMAKE="qmake-qt4"
 fi
 
-# bin/qmake 
+# bin/qmake
 if [ -f "/bin/qmake" ]; then
   QMAKE="qmake"
 fi
