+++ project/clean_coex.sh
@@ -1,4 +1,13 @@
-rm coex
-rm computer_examination/computer_examination
-rm computer_examination/Makefile
+if [ -f "coex" ]; then
+  rm coex
+fi
+
+if [ -f "computer_examination/computer_examination" ]; then
+  rm computer_examination/computer_examination
+fi
+
+if [ -f "computer_examination/Makefile" ]; then
+  rm computer_examination/Makefile
+fi
+
 rm computer_examination/*.o
